# schemas/openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"version": "1.0.0",
		"title": "Canine shelter API",
		"description": "An API for a canine shelter website",
		"contact": {
			"name": "Bartlomiej Wlodarski",
			"email": "wlodarsb@uni.coventry.ac.uk"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "https://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"servers": [
		{
			"url": "https://canoe-janet-3000.codio-box.uk/api/v1",
			"description": "API backend server"
		}
	],
	"security": [
		{
			"basic": [],
			"jwt": []
		}
	],
...
	"components": {
		"responses": {
			"authenticationError": {
				"description": "User needs to be authenticated to access this route"
			},
			"notFoundError": {
				"description": "Resource does not exist",
				"content": {
					"application/json": {
						"schema": {
							"type": "object",
							"properties": {
								"message": {
									"description": "Error message",
									"type": "string",
									"example": "User does not exist"
								}
							}
						}
					}
				}
			},
			"validationError": {
				"description": "Occurs when the server rejects the form data",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "./errors.json#/definitions/validationErrors"
						}
					}
				}
			}
		},
		"parameters": {
...
			"staffId": {
				"name": "id",
				"in": "path",
				"description": "Staff ID",
				"required": true,
				"schema": { "type": "integer" }
			},
			"uuid": {
				"name": "uuid",
				"in": "path",
				"description": "Image UUID",
				"required": true,
				"schema": { "type": "string", "format": "uuid" }
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"basic": {
				"type": "http",
				"scheme": "basic"
			}
		}
	},
	"paths": {
...
		"/staff": {
			"description": "A list of staff members",
			"get": {
				"summary": "Get a list of staff members",
				"description": "Gets a list of staff members, represented by a user ID and location ID they work at",
				"tags": ["Staff"],
				"responses": {
					"200": {
						"description": "Successfully retrieved list of staff members",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./staff.json#/definitions/staffList"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Only staff may view the staff list"
					}
				}
			},
			"post": {
				"tags": ["Staff"],
				"summary": "Regsiter a new staff member",
				"description": "Register a user as a staff member by assigning them to a location",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "./staff.json#/definitions/staff"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Successfully registered staff member",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./staff.json#/definitions/staffCreated"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/validationError"
					},
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Only users with the Staff role may set their location"
					}
				}
			}
		},
		"/staff/{id}": {
			"parameters": [
				{
					"$ref": "#/components/parameters/staffId"
				}
			],
			"summary": "A single staff member",
			"description": "Routes for managing individual staff members",
			"get": {
				"tags": ["Staff"],
				"summary": "Get a single staff member",
				"description": "Gets a single staff member by their staff ID",
				"responses": {
					"200": {
						"description": "Successfully retrieved staff member information",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./staff.json#/definitions/staffData"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Only staff may view staff records."
					},
					"404": {
						"$ref": "#/components/responses/notFoundError"
					}
				}
			},
			"put": {
				"tags": ["Staff"],
				"summary": "Update a staff member",
				"description": "Updates the staff member's data, such as their shelter location",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "./staff.json#/definitions/staff"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successfully updated staff record",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./staff.json#/definitions/staffUpdated"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/validationError"
					},
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Staff members may only update their own records"
					},
					"404": {
						"$ref": "#/components/responses/notFoundError"
					}
				}
			},
			"delete": {
				"summary": "Deletes a staff member",
				"description": "Deletes a staff member by staff ID.",
				"tags": ["Staff"],
				"responses": {
					"200": {
						"description": "Successfully deleted staff record",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./staff.json#/definitions/staffDeleted"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Staff may only delete their own records."
					},
					"404": {
						"$ref": "#/components/responses/notFoundError"
					}
				}
			}
		},
...
		"/auth/staff": {
			"summary": "User staff data",
			"description": "Route for getting user's staff data",
			"get": {
				"summary": "Gets the current user's staff record",
				"tags": ["Auth"],
				"responses": {
					"200": {
						"description": "Successfully retrieved staff record",
						"content": {
							"application/json": {
								"schema": { "$ref": "./staff.json#/definitions/staffData" }
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "User is not a staff member"
					}
				}
			}
		},
		"/uploads": {
			"summary": "Image uploads",
			"description": "Route for uploading images",
			"post": {
				"summary": "Uploads an image to the DB",
				"tags": ["Uploads"],
				"requestBody": {
					"description": "Image to upload",
					"required": true,
					"content": {
						"image/png": {
							"schema": {
								"type": "string",
								"format": "binary"
							}
						},
						"image/jpeg": {
							"schema": {
								"type": "string",
								"format": "binary"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successfully uploaded image",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./uploads.json#/definitions/imageAdded"
								}
							}
						}
					}
				}
			}
		},
		"/uploads/{uuid}": {
			"summary": "A single image",
			"description": "Route for getting an image from the DB",
			"parameters": [{ "$ref": "#/components/parameters/uuid" }],
			"get": {
				"summary": "Gets a single image from the DB",
				"tags": ["Uploads"],
				"security": [],
				"responses": {
					"200": {
						"description": "Successfully retrieved image",
						"content": {
							"image/png": {
								"schema": {
									"$ref": "./uploads.json#/definitions/image"
								}
							},
							"image/jpeg": {
								"schema": {
									"$ref": "./uploads.json#/definitions/image"
								}
							}
						}
					}
				}
			}
		}
	}
}